Updated based on code review suggestions.
